Database table and procedure changes

+++ LibraryManagement/dbo/Tables/Books.sql
@@ -2,7 +2,7 @@
     [Id]            NVARCHAR (50)  NOT NULL,
     [BookName]      NVARCHAR (MAX) NULL,
     [AuthorName]    NVARCHAR (MAX) NULL,
-    [Contact]       INT            NULL,
+    [Contact]       BIGINT         NULL,
     [Price]         MONEY          NULL,
     [PublishedDate] DATE           NULL,
     [Edition]       INT            NULL,
@@ -11,4 +11,6 @@
     [Copies]        INT            NULL,
     PRIMARY KEY CLUSTERED ([Id] ASC)
 );
+
+
